Remove client_count in favor of generic
